Update Pull 4

+++ src/map/packets/pet_sync.cpp
@@ -48,8 +48,8 @@ CPetSyncPacket::CPetSyncPacket(CCharEntity* PChar)
 
 	WBUFW(data,(0x0C)-4) = PChar->PPet->targid;
 	WBUFB(data,(0x0E)-4) = PChar->PPet->GetHPP();
-  int16 petTp = ((int16)PChar->PPet->health.tp) * 3.35  * 3;
-  WBUFW(data,(0x10)-4) = petTp;
+	int16 petTp = (PChar->PPet->health.tp) * 3.35  * 3;
+	WBUFW(data,(0x10)-4) = petTp;
 
 	// 0x14 - начинается имя питомца, но мы его записывать не будем, "мы экономить будем" © Матроскин
 }
